Implement createWeatherRecord function and update routing for weather data creation
