+++ weather-app-backend(API)/controllers/weatherController.js
@@ -1,22 +1,43 @@
-exports.getWeatherData = async (req, res) => {
-  const { lat, lon } = req.query;
+const Weather = require('../models/WeatherSchema');
 
-  if (!lat && !lon) {
+const createWeatherRecord = async (req, res) => {
+  const { lat, lon } = req.body;
+
+  if (!lat || !lon) {
     return res
       .status(400)
       .json({ error: 'Latitude and Longitude are required' });
   }
 
   try {
+    // Fetch current weather from OpenWeather API
     const apiKey = process.env.OPENWEATHER_API_KEY;
-    const url = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=${apiKey}`;
+    const weatherUrl = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=${apiKey}`;
 
-    const response = await fetch(url);
-    if (!response.ok) throw new Error('Failed to fetch weather data');
+    const weatherResponse = await fetch(weatherUrl); // Using Node.js 20 native fetch
+    if (!weatherResponse.ok) {
+      throw new Error('Failed to fetch weather data from OpenWeather API');
+    }
 
-    const weatherData = await response.json();
-    res.status(200).json(weatherData);
+    const weatherData = await weatherResponse.json();
+
+    // Extract relevant data
+    const newRecord = {
+      location: weatherData.name || 'Unknown Location',
+      lat,
+      lon,
+      temperature: weatherData.main.temp,
+      condition: weatherData.weather[0].description,
+    };
+
+    // Save the record to the database
+    const weather = new Weather(newRecord);
+    await weather.save();
+
+    res.status(201).json(weather); // Respond with the created record
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
 };
+
+module.exports = { createWeatherRecord };
